requirements.txt
Django==3.2
djangorestframework
django-cors-headers
Pillow==9.5.0